+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Change file system commit time from 30 seconds to 600 seconds.
 - Add support for using the Broadcom proprietary network driver for PCI Ethernet devices and USB adapters.
 - Add paccache timer for automatically cleaning up old packages.
+- Change Linux LTS to be based on Linux 6.12.
 
 ## 4.2.0 - The Ventoy Update